[ServiceProvider] Merged SafeTwigServiceProvider into TwigServiceProvider
