+++ src/Application.php
@@ -192,7 +192,6 @@ class Application extends Silex\Application
     public function initRendering()
     {
         $this->register(new Provider\TwigServiceProvider());
-        $this->register(new Provider\SafeTwigServiceProvider());
 
         $this->register(new Provider\RenderServiceProvider());
         $this->register(new Provider\RenderServiceProvider(true));
